PROX-396: add cds-proto
