+++ pom.xml
@@ -13,7 +13,7 @@
 
     <groupId>com.rbkmoney.adapter-thrift-lib</groupId>
     <artifactId>adapter-thrift-lib</artifactId>
-    <version>1.0.0</version>
+    <version>1.0.1</version>
     <packaging>pom</packaging>
 
     <properties>
@@ -31,6 +31,7 @@
 
     <modules>
         <module>damsel-utils</module>
+        <module>cds-utils</module>
     </modules>
 
     <build>
